pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rank1-factorization"
version = "1.0.0"
description = "Rank-1 factorization tests for particle physics datasets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Research Software Engineering", email = "research@example.org" }
]
keywords = ["particle physics", "factorization", "rank test", "HEP", "statistics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "matplotlib>=3.7",
    "requests>=2.28",
    "tqdm>=4.65",
    "typer>=0.9",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "uproot>=5.0",
    "awkward>=2.0",
    "pyarrow>=12.0",
    "rich>=13.0",
]

[project.optional-dependencies]
pdf = [
    "camelot-py[base]>=0.11",
    "pdfplumber>=0.10",
    "ghostscript>=0.7",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.3",
    "ruff>=0.1",
    "black>=23.0",
    "mypy>=1.5",
    "types-requests>=2.28",
    "types-PyYAML>=6.0",
    "types-tqdm>=4.65",
]
all = [
    "rank1-factorization[pdf,dev]",
]

[project.scripts]
rank1 = "rank1.cli:app"

[project.urls]
Homepage = "https://github.com/research/rank1-factorization"
Documentation = "https://github.com/research/rank1-factorization#readme"
Repository = "https://github.com/research/rank1-factorization.git"

[tool.hatch.build.targets.wheel]
packages = ["src/rank1"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["rank1"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests that require network access",
]
